% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allocation_funcs.R
\name{AllocM}
\alias{AllocM}
\title{Internal allocation M function}
\usage{
AllocM(bhgData, n0)
}
\arguments{
\item{bhgData}{Data giving the average number of price observations in the stratum for group.}

\item{n0}{Allocation}
}
\value{
m
}
\description{
Not sure if actually in use in price index
}
