% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allocation_funcs.R
\name{AllocTot}
\alias{AllocTot}
\title{Internal allocation function}
\usage{
AllocTot(s2, w, ai, totn)
}
\arguments{
\item{s2}{Sigma squared values}

\item{w}{Weights vector}

\item{ai}{Adjustment factor}

\item{totn}{Total sample size}
}
\value{
ni
}
\description{
Not sure if used in price index ...
}
