% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_funcs.R
\name{CalcInd}
\alias{CalcInd}
\title{Calculation of the estimate for a price index}
\usage{
CalcInd(data, baseVar, pVar, type, groupVar, consumVar = NULL, wVar = NULL)
}
\arguments{
\item{data}{The dataset}

\item{baseVar}{The variable name for the base goods price.}

\item{pVar}{The variable name for the goods price.}

\item{type}{The type of index to calculate. Choose from: 'carli', 'dutot', 'jevons'.}

\item{groupVar}{The variable used for grouping observations - sometimes called the elementary group}

\item{consumVar}{The variable used for grouping elementary groups. This may be the publishing level.}

\item{wVar}{The variable used to weight the elementary groups up to the consumer publishing level.}
}
\value{
The estimates for the index are returned as a vector of length equal to the number of groups (in groupVar or consumVar if specified).
}
\description{
Calculation of a price index
}
\examples{
{
data(priceData)
# Calculate index for consumer groups
CalcInd(data = priceData, baseVar = "b1", pVar = "p1", groupVar = "varenr", wVar = "weight", 
  consumVar = "coicop", type = "dutot")
CalcInd(data = priceData, baseVar = "b1", pVar = "p1", groupVar = "varenr", wVar = "weight", 
  consumVar = "coicop", type = "carli")
CalcInd(data = priceData, baseVar = "b1", pVar = "p1", groupVar = "varenr", wVar = "weight", 
  consumVar = "coicop", type = "jevons")
  
# Calculate index for elemenatry groups (weighted)
CalcInd(data = priceData, baseVar = "b1", pVar = "p1", groupVar = "varenr", wVar = "weight", 
   type = "dutot")
   
   }
}
