% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_funcs.R
\name{CalcIndS2}
\alias{CalcIndS2}
\title{Calculation of variance/sigma squared for price index}
\usage{
CalcIndS2(data, baseVar, pVar, groupVar, type = c("carli", "dutot", "jevons"))
}
\arguments{
\item{data}{The dataset}

\item{baseVar}{The variable name for the base goods price.}

\item{pVar}{The variable name for the goods price.}

\item{groupVar}{The variable used for grouping observations - sometimes called the elementary group}

\item{type}{The type of index to calculate. Choose from: 'carli', 'dutot', 'jevons'}
}
\value{
\item{s2}{sigma squred values for each strata group}
  \item{ai}{adjustment factor used in calculations} 
  \item{zmat}{Matrix of z values which can be used to check for outliers}
}
\description{
Calculation of sigma squared for a price index.
}
\examples{
{
# Call in test dataset for an index
data(priceData)

# Calculate s2 for an index
ss <- CalcIndS2(data = priceData, baseVar = "b1", pVar = "p1", groupVar = "varenr", 
type = "carli")
ss <- CalcIndS2(data = priceData, baseVar = "b1", pVar = "p1", groupVar = "coicop", 
type = "jevons")

# print out s2 values
ss$s2
}

}
