% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_funcs.R
\name{Carli}
\alias{Carli}
\title{Carli index calculation}
\usage{
Carli(x, y, na.rm = TRUE, ze.rm = TRUE)
}
\arguments{
\item{x}{Vector of base prices.}

\item{y}{Vector of prices in the statistical time period.}

\item{na.rm}{Logical on whether to remove missing values for x or y in the calculation.}

\item{ze.rm}{Logical on whether to remove missing values for the instrument variable for the calculations.}
}
\value{
Price index (L-type)
}
\description{
Carli index calculation
}
