% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_funcs.R
\name{Jevons}
\alias{Jevons}
\title{Jevons index caluclation}
\usage{
Jevons(x, y)
}
\arguments{
\item{x}{Vector of base prices.}

\item{y}{Vector of prices in the statistical time period.}
}
\value{
Price index (L-type)
}
\description{
Jevons index caluclation
}
