% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allocation_funcs.R
\name{SpecialRound}
\alias{SpecialRound}
\title{Rounding strata sample sizes}
\usage{
SpecialRound(x)
}
\arguments{
\item{x}{vector of unrounded strata sample sizes}
}
\value{
vector with round strata sizes where the total sample size is held
constant.
}
\description{
Function for rounding strata sample sizes to ensure the total sample size
remains the same
}
\examples{
{
 
jj <- c(4.2, 4.2, 4.2, 4.3, 4.3, 4.4, 4.4)
SpecialRound(jj)
}
}
